Fix MSVC C4244 warnings in version_converter adapters
Add explicit casts to resolve implicit double-to-float and int-to-float
conversions that become errors with ONNX_WERROR=ON after removing the
/wd4244 suppression flag.

- pad_10_11.h: static_cast<float>(node->f(kvalue))
- resize_10_11.h: use 0.0f and 1.0f float literals
- dropout_11_12.h: static_cast<float>(node->f(kratio)) and 0.5f
- clip_10_11.h: static_cast<float>(node->f(kmin/kmax))


Agent-signed-off: Developer (abdfea8a) [claude-opus-4.7 via copilot]

+++ onnx/version_converter/adapters/pad_10_11.h
@@ -37,7 +37,7 @@ class Pad_10_11 final : public Adapter {
       Tensor t_value;
       t_value.elem_type() = TensorProto_DataType_FLOAT;
       auto& data_value = t_value.floats();
-      data_value.emplace_back(node->f(kvalue));
+      data_value.emplace_back(static_cast<float>(node->f(kvalue)));
       Node* constant = graph->create(kConstant);
       constant->insertBefore(node);
       constant->t_(kvalue, t_value);
